Add Mesos and Python egg download page
We were previously providing an index of all the downloadable Mesos deb,
rpm, and native python egg packages via Jekyll. Now that the main
website is hosted on Wordpress we don't have a good mechanism for
generating this content.

This adds download links to the "open" website so that we can easily
generate indexes when new releases are available. In addition to this
patch, we will need to update the main website so that people can find
links to the direct deb, rpm, and egg downloads.